third commit - bugs

+++ src/particle_filter.h
@@ -36,7 +36,7 @@ class ParticleFilter {
 	bool is_initialized;
 	
 	// Vector of weights of all particles
-	std::vector<double> weights;
+	std::vector<float> weights;
 	
 public:
